trading-api-repo/infra/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-eks
cdk-ec2-key-pair==2.2.0
cdk8s-aws-load-balancer-controller
aws-cdk.aws-kms
aws-cdk.aws-lambda
pyyaml
requests
boto3
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudfront-origins==1.111.0
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-autoscaling
aws-cdk.aws-rds
aws-cdk.aws-secretsmanager
aws-cdk.aws-lambda
aws-cdk.aws-events-targets
aws-cdk.aws-cloudwatch
aws-cdk.aws-elasticsearch
aws-cdk.aws_elasticache
aws-cdk.aws_docdb
